requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
jsonschema==3.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
mssql-django==1.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
ofxtools==0.9.4
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
